pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Cur1yB <alexandr.ignatev1996@gmail.com>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Intended Audience :: Education",
    "Topic :: Games/Entertainment"
]

[tool.poetry.dependencies]
python = "^3.11.9"
pyyaml = "^6.0.1"
flask = "^2.0"
gunicorn = "^20.1.0"
python-dotenv = "^0.20.0"
[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
flake8 = "^7.0.0"
psycopg2 = "^2.9.3"
validators = "^0.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "page_analyzer.app:main"
